Added Java comments and modified some classes names.

+++ src/main/java/com/gomeetdr/service/DoctorService.java
@@ -1,8 +1,10 @@
 package com.gomeetdr.service;
- 
+
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 
 import com.gomeetdr.modal.Appointment;
@@ -15,32 +17,41 @@ public class DoctorService {
 
 	@Autowired
 	private DoctorRepository doctorRepository;
-	
+
 	public Doctor createOrUpdateDoctor(Doctor doctor) {
 		return doctorRepository.save(doctor);
 	}
-	
+
 	public Doctor updateDoctor(Doctor doctor) throws NotFoundException {
-		if(doctorRepository.exists(doctor.getId())) {
+		if (doctorRepository.exists(doctor.getId())) {
 			return doctorRepository.save(doctor);
 		}
-		
-		throw new NotFoundException("Doctor not found");
-	}
-	
-	public Doctor getDoctor(Long id) {
-		return doctorRepository.findOne(id);
-	}
-	
-	public Iterable<Doctor> getAllDoctors() {
-		return doctorRepository.findAll();
-	}
-	
-	public void deleteDoctor(Long id) {
-		doctorRepository.delete(id);
+		throw new NotFoundException("Could not update doctor because given id does not exists");
 	}
 
-	public List<Appointment> GetAppointments(Long id) {
-		return doctorRepository.findOne(id).getAppointments();
+	public Doctor getDoctor(Long id) throws NotFoundException {
+		if (doctorRepository.exists(id)) {
+			return doctorRepository.findOne(id);
+		}
+		throw new NotFoundException("Could not get doctor because given id does not exists");
+	}
+
+	public Iterable<Doctor> getAllDoctors() {
+		Sort sort = new Sort(new Sort.Order(Direction.ASC, "id"));
+		return doctorRepository.findAll(sort);
+	}
+
+	public void deleteDoctor(Long id) throws NotFoundException {
+		if (doctorRepository.exists(id)) {
+			doctorRepository.delete(id);
+		}
+		throw new NotFoundException("Could not delete doctor because given id does not exists.");
+	}
+
+	public List<Appointment> GetAppointments(Long id) throws NotFoundException {
+		if (doctorRepository.exists(id)) {
+			return doctorRepository.findOne(id).getAppointments();
+		}
+		throw new NotFoundException("Could not get appointments because given id does not exists.");
 	}
 }
